+++ packages/core/lib/combo/confvars.ts
@@ -210,6 +210,8 @@ export const CONFVARS = [
   'SHARED_HAMMER',
   'MM_UPGRADES_STICKS_NUTS',
   'OOT_SHUFFLE_EGGS',
+  'MM_STONE_OF_AGONY',
+  'SHARED_STONE_OF_AGONY',
 ] as const;
 
 export type Confvar = typeof CONFVARS[number];